refactor: Extract `Team` class
Create a class representing a team which has a seed and region. There
were a lot of places where we had parallel variables to track both.
Creating a single class to hold both attributes means we can eliminate
half of those variables, and we don't need any logic to correlate seeds
and regions since they are encapsulated under the same reference now.